// R1.2_Documents/51_製造/21_参考資料/HTML_FR-CFM-DFM_起動手順/HTML_FR-CFM-DFM/source/src/app/types/flm/car.ts
export interface TerminalChangeParams {
  car: {
    komtrax_unit: {
      terminal_component: {
        serial?: string;
        part?: string;
      };
    };
    car_identification: {
      initial_smr: string;
      update_datetime: string;
    };
  };
}
// ...
export interface TerminalChangeDisplayData {
  common: {
    car_identification: {
      maker_name: string;
      models: string;
      type_revs: string;
      serials: string;
    };
  };
  car: {
    support_distributor: {
      label?: string;
    };
  };
  modem: {
    serial: string;
    part: string;
  };
}